Finalize migration to loading Config from Spec instead of CliArgs by moving and reworking existing test cases
