+++ detekt-cli/src/test/kotlin/io/gitlab/arturbosch/detekt/cli/TestFactory.kt
@@ -11,4 +11,6 @@ fun createCliArgs(vararg args: String): CliArgs {
     return parseArguments(args, NullPrintStream(), NullPrintStream())
 }
 
+fun CliArgs.toSpec() = createSpec(NullPrintStream(), NullPrintStream())
+
 fun createRunner(cliArgs: CliArgs): Runner = Runner(cliArgs, NullPrintStream(), NullPrintStream())
