+++ detekt-core/src/main/kotlin/io/gitlab/arturbosch/detekt/core/config/Configurations.kt
