+++ detekt-core/src/test/resources/configs/config-with-auto-correct.yml
