+++ detekt-core/src/test/resources/configs/fail-fast-will-override-here.yml
