+++ detekt-core/src/test/resources/configs/fail-fast-wont-override-here.yml
