+++ detekt-core/src/test/resources/configs/one.yml
