+++ detekt-core/src/test/resources/configs/three.yml
