+++ detekt-core/src/test/resources/configs/two.yml
