+++ detekt-test-utils/src/main/kotlin/io/github/detekt/test/utils/Resources.kt
@@ -1,18 +1,21 @@
 package io.github.detekt.test.utils
 
 import java.net.URI
+import java.net.URL
 import java.nio.file.Files
 import java.nio.file.Path
 import java.nio.file.Paths
 
 internal object Resources
 
-fun resource(name: String): URI {
+fun resourceUrl(name: String): URL {
     val explicitName = if (name.startsWith("/")) name else "/$name"
     val resource = Resources::class.java.getResource(explicitName)
     requireNotNull(resource) { "Make sure the resource '$name' exists!" }
-    return resource.toURI()
+    return resource
 }
+
+fun resource(name: String): URI = resourceUrl(name).toURI()
 
 fun resourceAsPath(name: String): Path = Paths.get(resource(name))
